nicole: init: Set RPS for rmnet when modem is up
Stock does this based on gsm.sim.quectel.boot_completed.
However, that is set from prebuilt TeleService. From observation, this
prop comes online as soon as the modem is up, thus it seems to be a
good candidate for accurate execution without having to modify src or
include the prebuilt.

Signed-off-by: AnierinB <anierin@evolution-x.org>